Almost ready for prime time.

+++ quickpwn.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import os.path
 
 from autopwn_suite.api import AutoScanner
 import threading
@@ -19,12 +20,17 @@ debug_scan = True
 target_ports = [20, 21, 22, 23, 25, 43, 53, 80, 110, 123, 137, 138, 139, 143, 161, 162, 389, 443, 445, 500, 554, 587,
                 993, 1434, 3306, 3389, 8008, 8080, 5900]
 ip_list_string = ""
+all_scan_results = {}
+results_dir = "results/"
+results_partial_dir = "results/partial/"
+results_full_dir = "results/full/"
 
 
 # nmap -oX - 192.168.50.129 -sV -T 5 -O
 
 def scan_host(ip: str, do_full_scan=False):
-    global host_timeout, scan_speed, api_key, os_scan, scan_vulns, nmap_args, debug_scan, target_ports
+    global host_timeout, scan_speed, api_key, os_scan, scan_vulns, nmap_args, debug_scan, target_ports, all_scan_results
+    global results_partial_dir, results_full_dir
 
     print("Starting scan.")
     scanner = AutoScanner()
@@ -43,9 +49,12 @@ def scan_host(ip: str, do_full_scan=False):
                                os_scan=os_scan, scan_vulns=scan_vulns, nmap_args=full_nmap_args, debug=debug_scan)
 
         if do_full_scan:
-            scanner.save_to_file(ip + "_full_results.json")
+            scanner.save_to_file(results_full_dir + ip + "_full_results.json")
         else:
-            scanner.save_to_file(ip + "_partial_results.json")
+            scanner.save_to_file(results_partial_dir + ip + "_partial_results.json")
+
+        for ip in results:
+            all_scan_results[ip] = results[ip]
     except Exception as e:
         print(e)
         print(e.args)
@@ -84,10 +93,19 @@ def parse_scan_results(scan_result):
     exit(0)
 
 
+def wait_for_thread_list_to_end(threadlist: list):
+    while len(threadlist) > 0:
+        time.sleep(1)
+        for thread in threadlist:
+            if not thread.is_alive():
+                threadlist.remove(thread)
+
+
 def main():
     global api_key
     global subnet
     global ip_list_string
+    global all_scan_results
     scan_thread_list = []
     ping_thread_list = []
 
@@ -100,35 +118,33 @@ def main():
     except:
         print("Unable to load api key.")
 
-    try:
+    if os.path.isfile("IPS_FIRING_RANGE.TXT"):
         with open("IPS_FIRING_RANGE.TXT", "r") as ips:
             for ip in ips:
                 ip = ip.strip()
                 new_thread = threading.Thread(target=scan_host, args=[ip, True])
-                # scan_host(ip)
                 new_thread.start()
                 scan_thread_list.append(new_thread)
-    except:
+
+        wait_for_thread_list_to_end(scan_thread_list)
+
+        with open(results_dir + "full_results.json", "w") as outfile:
+            json.dump(all_scan_results, outfile, indent=2)
+
+    else:
         for i in range(0, 256):
             ip = subnet + str(i)
             new_thread = threading.Thread(target=ping_host, args=[ip])
             new_thread.start()
             ping_thread_list.append(new_thread)
 
-        while len(ping_thread_list) > 0:
-            time.sleep(1)
-            for thread in ping_thread_list:
-                if not thread.is_alive():
-                    ping_thread_list.remove(thread)
+        wait_for_thread_list_to_end(ping_thread_list)
 
         with open("IPS_FIRING_RANGE.TXT", "w") as outfile:
             outfile.write(ip_list_string)
 
-    while len(scan_thread_list) > 0:
-        time.sleep(1)
-        for thread in scan_thread_list:
-            if not thread.is_alive():
-                scan_thread_list.remove(thread)
+        with open(results_dir + "partial_results.json", "w") as outfile:
+            json.dump(all_scan_results, outfile, indent=2)
 
 
 if __name__ == "__main__":
